switch x12 to use unsigned ticks
 * the clock used for system ticks is 32-bit.  it takes a minute for it
 to overflow (where overflow means counting 2^32 times, not simply
 wrapping to 0).  we dont care if it wraps to 0 as unsigned math will
 still give the correct difference value even after wrapping.

+++ cubeide_proj/Core/SwitecX12-lib/SwitecX12.hpp
@@ -15,7 +15,7 @@ public:
   uint32_t steps;            // total steps available
   long time0;           // time when we entered this state
   unsigned int microDelay;       // microsecs until next state
-  unsigned short (*accelTable)[2]; // accel table can be modified.
+  uint32_t (*accelTable)[2]; // accel table can be modified.
   int maxVel;           // fastest vel allowed
   int vel;              // steps travelled under acceleration
   int dir;                      // direction -1,0,1
